requirements/base.txt
django==3.0.2
django-environ==0.4.5
qiniu==7.2.6
mysqlclient==1.4.6
django-ckeditor==5.6.1
Pillow==7.0.0
django-mptt==0.10.0
django-extensions==2.2.5

# rest framework
django-filter==2.2.0
djangorestframework==3.11.0
django-crispy-forms==1.8.1
# deploy
gunicorn==20.0.4
gevent==1.4.0